// src-ui/core/ipc/generated/debug/DebugChannelChoice.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DebugChannelId } from "./DebugChannelId";

/**
 * 某个 Bot 上一次选的调用通道和事件通道
 */
export type DebugChannelChoice = { call: DebugChannelId, events: DebugChannelId, };