// src-ui/core/ipc/generated/debug/DebugLayout.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DebugChatView } from "./DebugChatView";

export type DebugLayout = { left_collapsed: boolean, right_collapsed: boolean, left_width: number, right_width: number, right_view: DebugChatView, };