// src-ui/core/ipc/generated/debug/DebugWorkspace.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DebugChannelChoice } from "./DebugChannelChoice";
import type { DebugLayout } from "./DebugLayout";
import type { DebugRequestDraft } from "./DebugRequestDraft";

/**
 * 调试台的工作区：重开应用后回到离开时的样子
 */
export type DebugWorkspace = { version: number, tabs: Array<DebugRequestDraft>, active_tab: string | null, 
/**
 * 最近关掉的 10 个标签，Ctrl+Shift+T 用
 */
closed_tabs: Array<DebugRequestDraft>, selected_bot: string | null, 
/**
 * 以 `bot_id` 为键
 */
channel_choice: { [key in string]: DebugChannelChoice }, layout: DebugLayout, 
/**
 * 最近用过的 20 个动作名
 */
recent_actions: Array<string>, };